Move around actions to more relevant controllers
