+++ lib/api/controllers/authController.js
@@ -43,7 +43,7 @@ function AuthController (kuzzle) {
   this.login = function authLogin (request) {
     var strategy;
 
-    assertBody(request, 'login');
+    assertBody(request, 'auth:login');
     strategy = request.input.body.strategy || 'local';
 
     return kuzzle.passport.authenticate({query: request.input.body}, strategy)
@@ -110,8 +110,8 @@ function AuthController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.checkToken = function authCheckToken (request) {
-    assertBody(request, 'checkToken');
-    assertBodyAttribute(request, 'token', 'checkToken');
+    assertBody(request, 'auth:checkToken');
+    assertBodyAttribute(request, 'token', 'auth:checkToken');
 
     return kuzzle.repositories.token.verifyToken(request.input.body.token)
       .then(token => Promise.resolve({valid: true, expiresAt: token.expiresAt}))
