+++ lib/api/controllers/bulkController.js
@@ -5,7 +5,11 @@ var
   assertBody = require('./util/requestAssertions').assertBody,
   assertBodyAttribute = require('./util/requestAssertions').assertBodyAttribute;
 
-module.exports = function BulkController (kuzzle) {
+/**
+ * @param {Kuzzle} kuzzle
+ * @constructor
+ */
+function BulkController (kuzzle) {
   var engine = kuzzle.services.list.storageEngine;
 
   /**
@@ -15,8 +19,8 @@ module.exports = function BulkController (kuzzle) {
    * @returns {Promise}
    */
   this.import = function bulkImport (request) {
-    assertBody(request, 'import');
-    assertBodyAttribute(request, 'bulkData', 'import');
+    assertBody(request, 'bulk:import');
+    assertBodyAttribute(request, 'bulkData', 'bulk:import');
 
     return engine.import(request)
       .then(response => {
@@ -27,5 +31,6 @@ module.exports = function BulkController (kuzzle) {
         return Promise.resolve(response);
       });
   };
+}
 
-};
+module.exports = BulkController;
