+++ lib/api/controllers/cli/data.js
@@ -15,7 +15,7 @@ var
 function data (request) {
   var promises = [];
 
-  assertBody(request, 'data');
+  assertBody(request, 'cli:data');
 
   if (request.input.body.fixtures) {
     Object.keys(request.input.body.fixtures).forEach((index => {
