+++ lib/api/controllers/cliController.js
@@ -11,8 +11,8 @@ function CliController (kuzzle) {
 
   this.init = () => {
     this.actions = {
-      adminExists: request => kuzzle.funnel.controllers.admin.adminExists(request),
-      createFirstAdmin: request => kuzzle.funnel.controllers.admin.createFirstAdmin(request),
+      adminExists: request => kuzzle.funnel.controllers.server.adminExists(request),
+      createFirstAdmin: request => kuzzle.funnel.controllers.security.createFirstAdmin(request),
       cleanDb: require('./cli/cleanDb')(kuzzle),
       clearCache: require('./cli/clearCache')(kuzzle),
       dump:require('./cli/dump')(kuzzle),
