+++ lib/api/controllers/documentController.js
@@ -1,19 +1,65 @@
 'use strict';
 
 var
+  Promise = require('bluebird'),
+  BadRequestError = require('kuzzle-common-objects').errors.BadRequestError,
   assertBody = require('./util/requestAssertions').assertBody,
   assertId = require('./util/requestAssertions').assertId,
-  assertIndex = require('./util/requestAssertions').assertIndex,
-  assertIndexAndCollection = require('./util/requestAssertions').assertIndexAndCollection,
   assertBodyAttribute = require('./util/requestAssertions').assertBodyAttribute,
-  BadRequestError = require('kuzzle-common-objects').errors.BadRequestError;
+  assertIndexAndCollection = require('./util/requestAssertions').assertIndexAndCollection;
 
 /**
  * @param {Kuzzle} kuzzle
  * @constructor
  */
-function WriteController (kuzzle) {
+function DocumentController (kuzzle) {
   var engine = kuzzle.services.list.storageEngine;
+
+
+  /**
+   * @param {Request} request
+   * @returns {Promise<Object>}
+   */
+  this.search = function documentSearch (request) {
+    assertIndexAndCollection(request, 'search');
+
+    if (request.input.resource.index.split(',').length > 1) {
+      throw new BadRequestError('search on multiple indexes is not available.');
+    }
+    if (request.input.resource.collection.split(',').length > 1) {
+      throw new BadRequestError('search on multiple collections is not available.');
+    }
+
+    return engine.search(request);
+  };
+
+  /**
+   * @param {Request} request
+   * @returns {Promise<Object>}
+   */
+  this.scroll = function documentScroll (request) {
+    return engine.scroll(request);
+  };
+
+  /**
+   * @param {Request} request
+   * @returns {Promise<Object>}
+   */
+  this.get = function documentGet (request) {
+    assertId(request, 'document:get');
+
+    return engine.get(request);
+  };
+
+  /**
+   * @param {Request} request
+   * @returns {Promise<Object>}
+   */
+  this.count = function documentCount (request) {
+    assertBody(request, 'document:count');
+
+    return engine.count(request);
+  };
 
   /**
    * Create a new document
@@ -21,12 +67,12 @@ function WriteController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.create = function writeCreate (request) {
+  this.create = function documentCreate (request) {
     /** @type KuzzleRequest */
     var modifiedRequest;
 
-    assertBody(request, 'create');
-    assertIndexAndCollection(request, 'create');
+    assertBody(request, 'document:create');
+    assertIndexAndCollection(request, 'document:create');
 
     return kuzzle.validation.validationPromise(request, false)
       .then(newRequest => {
@@ -45,32 +91,17 @@ function WriteController (kuzzle) {
   };
 
   /**
-   * Publish a realtime message
-   *
-   * @param {Request} request
-   * @returns {Promise<Object>}
-   */
-  this.publish = function writePublish (request) {
-
-    assertBody(request, 'publish');
-    assertIndexAndCollection(request, 'publish');
-
-    return kuzzle.validation.validationPromise(request, false)
-      .then(newRequest => kuzzle.notifier.publish(newRequest));
-  };
-
-  /**
    * Create a new document through the persistent layer. If it already exists, update it instead.
    *
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.createOrReplace = function writeCreateOrReplace (request) {
+  this.createOrReplace = function documentCreateOrReplace (request) {
     /** @type KuzzleRequest */
     var modifiedRequest;
 
-    assertBody(request, 'createOrReplace');
-    assertIndexAndCollection(request, 'createOrReplace');
+    assertBody(request, 'document:createOrReplace');
+    assertIndexAndCollection(request, 'document:createOrReplace');
 
     return kuzzle.validation.validationPromise(request, false)
       .then(newRequest => {
@@ -101,13 +132,13 @@ function WriteController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.update = function writeUpdate (request) {
+  this.update = function documentUpdate (request) {
     /** @type KuzzleRequest */
     var modifiedRequest;
 
-    assertBody(request, 'update');
-    assertIndexAndCollection(request, 'update');
-    assertId(request, 'update');
+    assertBody(request, 'document:update');
+    assertIndexAndCollection(request, 'document:update');
+    assertId(request, 'document:update');
 
     return kuzzle.validation.validationPromise(request, false)
       .then(newRequest => {
@@ -128,13 +159,13 @@ function WriteController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.replace = function writeReplace (request) {
+  this.replace = function documentReplace (request) {
     /** @type KuzzleRequest */
     var modifiedRequest;
 
-    assertBody(request, 'replace');
-    assertIndexAndCollection(request, 'replace');
-    assertId(request, 'replace');
+    assertBody(request, 'document:replace');
+    assertIndexAndCollection(request, 'document:replace');
+    assertId(request, 'document:replace');
 
     return kuzzle.validation.validationPromise(request, false)
       .then(newRequest => {
@@ -158,9 +189,9 @@ function WriteController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.delete = function writeDelete (request) {
-    assertIndexAndCollection(request, 'delete');
-    assertId(request, 'delete');
+  this.delete = function documentDelete (request) {
+    assertIndexAndCollection(request, 'document:delete');
+    assertId(request, 'document:delete');
 
     request.input.args.state = 'pending';
     kuzzle.notifier.publish(request);
@@ -179,10 +210,10 @@ function WriteController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.deleteByQuery = function writeDeleteByQuery (request) {
-    assertBody(request, 'deleteByQuery');
-    assertIndexAndCollection(request, 'deleteByQuery');
-    assertBodyAttribute(request, 'query', 'deleteByQuery');
+  this.deleteByQuery = function documentDeleteByQuery (request) {
+    assertBody(request, 'document:deleteByQuery');
+    assertIndexAndCollection(request, 'document:deleteByQuery');
+    assertBodyAttribute(request, 'query', 'document:deleteByQuery');
 
     if (!(request.input.body.query instanceof Object)) {
       throw new BadRequestError('deleteByQuery must specify a not empty body attribute "query" of type object.');
@@ -196,22 +227,6 @@ function WriteController (kuzzle) {
       });
   };
 
-  /**
-   * Creates a new collection. Does nothing if the collection already exists.
-   *
-   * @param {Request} request
-   * @returns {Promise<Object>}
-   */
-  this.createCollection = function writeCreateCollection (request) {
-    assertIndex(request, 'createCollection');
-
-    return engine.createCollection(request)
-      .then(response => {
-        kuzzle.indexCache.add(request.input.resource.index, request.input.resource.collection);
-
-        return Promise.resolve(response);
-      });
-  };
 
   /**
    * Validate a document against collection validation specifications.
@@ -219,10 +234,10 @@ function WriteController (kuzzle) {
    * @param {Request} request
    * @returns {Promise<Object>}
    */
-  this.validate = function writeValidate (request) {
-    assertBody(request, 'validate');
-    assertIndexAndCollection(request, 'validate');
-    assertId(request, 'validate');
+  this.validate = function documentValidate (request) {
+    assertBody(request, 'document:validate');
+    assertIndexAndCollection(request, 'document:validate');
+    assertId(request, 'document:validate');
 
     return kuzzle.validation.validationPromise(request, true)
       .then(response => {
@@ -235,4 +250,4 @@ function WriteController (kuzzle) {
   };
 }
 
-module.exports = WriteController;
+module.exports = DocumentController;
