+++ lib/api/controllers/funnelController.js
@@ -3,14 +3,15 @@
 var
   Promise = require('bluebird'),
   Deque = require('denque'),
-  MemoryStorageController = require('./memoryStorageController'),
-  WriteController = require('./writeController'),
-  ReadController = require('./readController'),
-  SubscribeController = require('./subscribeController'),
-  BulkController = require('./bulkController'),
-  AdminController = require('./adminController'),
-  SecurityController = require('./securityController'),
   AuthController = require('./authController'),
+  BulkController = require('./bulkController'),
+  CollectionController = require('./collectionController'),
+  DocumentController = require('./documentController'),
+  IndexController = require('./indexController'),
+  MemoryStorageController = require('./memoryStorageController'),
+  RealtimeController = require('./realtimeController'),
+  SecurityController = require('./securityController'),
+  ServerController = require('./serverController'),
   BadRequestError = require('kuzzle-common-objects').errors.BadRequestError,
   UnauthorizedError = require('kuzzle-common-objects').errors.UnauthorizedError,
   ForbiddenError = require('kuzzle-common-objects').errors.ForbiddenError,
@@ -50,14 +51,15 @@ function FunnelController (kuzzle) {
   this.historized = 0;
 
   this.init = function funnelInit () {
-    this.controllers.admin = new AdminController(kuzzle);
     this.controllers.auth = new AuthController(kuzzle);
     this.controllers.bulk = new BulkController(kuzzle);
+    this.controllers.collection = new CollectionController(kuzzle);
+    this.controllers.document = new DocumentController(kuzzle);
+    this.controllers.index = new IndexController(kuzzle);
     this.controllers.memoryStorage = this.controllers.ms = new MemoryStorageController(kuzzle);
-    this.controllers.read = new ReadController(kuzzle);
+    this.controllers.realtime = new RealtimeController(kuzzle);
     this.controllers.security = new SecurityController(kuzzle);
-    this.controllers.subscribe = new SubscribeController(kuzzle);
-    this.controllers.write = new WriteController(kuzzle);
+    this.controllers.server = new ServerController(kuzzle);
 
     kuzzle.pluginsManager.injectControllers();
   };
