+++ lib/api/controllers/securityController.js
@@ -25,7 +25,7 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.getRole = function securityGetRole (request) {
-    assertId(request, 'getRole');
+    assertId(request, 'security:getRole');
 
     return kuzzle.repositories.role.loadRole(request.input.resource._id)
       .then(role => {
@@ -44,8 +44,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.mGetRoles = function securityMGetRoles (request) {
-    assertBody(request, 'mGetRoles');
-    assertBodyAttribute(request, 'ids', 'mGetRoles');
+    assertBody(request, 'security:mGetRoles');
+    assertBodyAttribute(request, 'ids', 'security:mGetRoles');
 
     if (!Array.isArray(request.input.body.ids)) {
       throw new BadRequestError('mGetRoles must specify an array of "ids"');
@@ -85,8 +85,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.createOrReplaceRole = function securityCreateOrReplaceRole (request) {
-    assertBody(request, 'createOrReplaceRole');
-    assertId(request, 'createOrReplaceRole');
+    assertBody(request, 'security:createOrReplaceRole');
+    assertId(request, 'security:createOrReplaceRole');
 
     return createOrReplaceRole.call(kuzzle, request, {method: 'createOrReplace'})
       .then(role => Promise.resolve(formatProcessing.formatRoleForSerialization(role)));
@@ -115,7 +115,7 @@ function SecurityController (kuzzle) {
   this.deleteRole = function securityDeleteRole (request) {
     var role;
 
-    assertId(request, 'deleteRole');
+    assertId(request, 'security:deleteRole');
 
     role = kuzzle.repositories.role.getRoleFromRequest(request);
 
@@ -148,11 +148,11 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.mGetProfiles = function securityMGetProfiles (request) {
-    assertBody(request, 'mGetProfiles');
-    assertBodyAttribute(request, 'ids', 'mGetProfiles');
+    assertBody(request, 'security:mGetProfiles');
+    assertBodyAttribute(request, 'ids', 'security:mGetProfiles');
 
     if (!Array.isArray(request.input.body.ids)) {
-      throw new BadRequestError('mGetProfiles must specify an array of "ids"');
+      throw new BadRequestError('security:mGetProfiles must specify an array of "ids"');
     }
 
     return kuzzle.repositories.profile.loadMultiFromDatabase(request.input.body.ids)
@@ -170,8 +170,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.createOrReplaceProfile = function securityCreateOrReplaceProfile (request) {
-    assertBody(request, 'createOrReplaceProfile');
-    assertBody(request, 'policies', 'createOrReplaceProfile');
+    assertBody(request, 'security:createOrReplaceProfile');
+    assertBody(request, 'policies', 'security:createOrReplaceProfile');
 
     if (!_.isArray(request.input.body.policies)) {
       throw new BadRequestError('createOrReplaceProfile must specify an array of "policies"');
@@ -188,8 +188,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.createProfile = function securityCreateProfile (request) {
-    assertBody(request, 'createProfile');
-    assertBody(request, 'policies', 'createProfile');
+    assertBody(request, 'security:createProfile');
+    assertBody(request, 'policies', 'security:createProfile');
 
     if (!_.isArray(request.input.body.policies)) {
       throw new BadRequestError('createOrReplaceProfile must specify an array of "policies"');
@@ -206,7 +206,7 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.deleteProfile = function securityDeleteProfile (request) {
-    assertId(request, 'deleteProfile');
+    assertId(request, 'security:deleteProfile');
 
     return kuzzle.repositories.profile.buildProfileFromRequest(request)
       .then(profile => kuzzle.repositories.profile.deleteProfile(profile));
@@ -242,7 +242,7 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.getUser = function securityGetUser (request) {
-    assertId(request, 'getUser');
+    assertId(request, 'security:getUser');
 
     return kuzzle.repositories.user.load(request.input.resource._id)
       .then(user => {
@@ -261,7 +261,7 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.getProfileRights = function securityGetProfileRights (request) {
-    assertId(request, 'getProfileRights');
+    assertId(request, 'security:getProfileRights');
 
     return kuzzle.repositories.profile.loadProfile(request.input.resource._id)
       .then(profile => {
@@ -282,7 +282,7 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.getUserRights = function securityGetUserRights (request) {
-    assertId(request, 'getUserRights');
+    assertId(request, 'security:getUserRights');
 
     return kuzzle.repositories.user.load(request.input.resource._id)
       .then(user => {
@@ -328,7 +328,7 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.deleteUser = function securityDeleteUser (request) {
-    assertId(request, 'deleteUser');
+    assertId(request, 'security:deleteUser');
 
     return kuzzle.repositories.user.delete(request.input.resource._id)
       .then(() => Promise.resolve({ _id: request.input.resource._id}));
@@ -345,11 +345,11 @@ function SecurityController (kuzzle) {
       user = new User(),
       pojoUser;
 
-    assertBody(request, 'createUser');
-    assertBodyAttribute(request, 'profileIds', 'createUser');
+    assertBody(request, 'security:createUser');
+    assertBodyAttribute(request, 'profileIds', 'security:createUser');
 
     if (!Array.isArray(request.input.body.profileIds)) {
-      throw new BadRequestError('createUser must specify an array of "profileIds"');
+      throw new BadRequestError('security:createUser must specify an array of "profileIds"');
     }
 
     pojoUser = request.input.body;
@@ -371,8 +371,8 @@ function SecurityController (kuzzle) {
       user = new User(),
       pojoUser;
 
-    assertBody(request, 'createUser');
-    assertBodyAttributeAbsence(request, 'profileIds', 'createUser');
+    assertBody(request, 'security:createUser');
+    assertBodyAttributeAbsence(request, 'profileIds', 'security:createUser');
 
     pojoUser = request.input.body;
     pojoUser._id = request.input.resource._id || uuid.v4();
@@ -390,8 +390,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.updateUser = function securityUpdateUser (request) {
-    assertBody(request, 'updateUser');
-    assertId(request, 'updateUser');
+    assertBody(request, 'security:updateUser');
+    assertId(request, 'security:updateUser');
 
     return kuzzle.repositories.user.load(request.input.resource._id)
       .then(user => kuzzle.repositories.user.persist(_.extend(user, request.input.body), {database: {method: 'update'}}))
@@ -406,8 +406,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.updateProfile = function securityUpdateProfile (request) {
-    assertBody(request, 'updateProfile');
-    assertId(request, 'updateProfile');
+    assertBody(request, 'security:updateProfile');
+    assertId(request, 'security:updateProfile');
 
     return kuzzle.repositories.profile.loadProfile(request.input.resource._id)
       .then(profile => kuzzle.repositories.profile.validateAndSaveProfile(_.extend(profile, request.input.body), {method: 'update'}))
@@ -421,8 +421,8 @@ function SecurityController (kuzzle) {
    * @returns {Promise<Object>}
    */
   this.updateRole = function securityUpdateRole (request) {
-    assertBody(request, 'updateRole');
-    assertId(request, 'updateRole');
+    assertBody(request, 'security:updateRole');
+    assertId(request, 'security:updateRole');
 
     return kuzzle.repositories.role.loadRole(request.input.resource._id)
       .then(role => {
@@ -444,8 +444,8 @@ function SecurityController (kuzzle) {
   this.createOrReplaceUser = function securityCreateOrReplaceUser (request) {
     var user = new User();
 
-    assertBody(request, 'createOrReplaceUser');
-    assertBodyAttribute(request, 'profileIds', 'createOrReplaceUser');
+    assertBody(request, 'security:createOrReplaceUser');
+    assertBodyAttribute(request, 'profileIds', 'security:createOrReplaceUser');
 
     if (request.input.resource._id) {
       user._id = request.input.resource._id;
@@ -456,7 +456,45 @@ function SecurityController (kuzzle) {
       .then(modifiedUser => formatProcessing.formatUserForSerialization(kuzzle, modifiedUser))
       .then(modifiedUser => Promise.resolve(modifiedUser));
   };
+
+  /**
+   * Creates the first admin user if it does not already exist
+   *
+   * @param {Request} request
+   * @returns {Promise<Object>}
+   */
+  this.createFirstAdmin = function securityCreateFirstAdmin (request) {
+    var reset;
+
+    assertBody(request, 'security:createFirstAdmin');
+
+    reset = request.input.args.reset || false;
+
+    return kuzzle.funnel.controllers.server.adminExists()
+      .then(adminExists => {
+        delete request.input.args.reset;
+        request.input.body.profileIds = ['admin'];
+
+        if (adminExists.exists) {
+          return Promise.reject(new Error('admin user is already set'));
+        }
+
+        return kuzzle.funnel.controllers.security.createOrReplaceUser(request);
+      })
+      .then(response => {
+        if (reset) {
+          return resetRoles.call(kuzzle)
+            .then(() => resetProfiles.call(kuzzle))
+            .then(() => kuzzle.funnel.controllers.index.refreshInternalIndex())
+            .then(() => Promise.resolve(response));
+        }
+
+        return Promise.resolve(response);
+      });
+  };
 }
+
+module.exports = SecurityController;
 
 /**
  * @this Kuzzle
@@ -482,4 +520,28 @@ function createOrReplaceProfile (kuzzle, request, opts) {
     .then(hydratedProfile => kuzzle.repositories.profile.validateAndSaveProfile(hydratedProfile, opts));
 }
 
-module.exports = SecurityController;
+/**
+ * @this {Kuzzle}
+ * @returns {Promise.<*>}
+ */
+function resetRoles () {
+  var promises;
+
+  promises = ['admin', 'default', 'anonymous'].map(id => {
+    return this.internalEngine.createOrReplace('roles', id, this.config.security.standard.roles[id]);
+  });
+
+  return Promise.all(promises);
+}
+
+/**
+ * @this {Kuzzle}
+ * @returns {Promise.<*>}
+ */
+function resetProfiles () {
+  return this.internalEngine
+    .createOrReplace('profiles', 'admin', {policies: [{roleId: 'admin', allowInternalIndex: true}]})
+    .then(() => this.internalEngine.createOrReplace('profiles', 'anonymous', {policies: [{roleId: 'anonymous'}]}))
+    .then(() => this.internalEngine.createOrReplace('profiles', 'default', {policies: [{roleId: 'default'}]}));
+}
+
